Add a data source for AWS credential prerequisites.
This change adds a data source for credential prerequisites -- the account ID
and external ID needed to create delegated access roles on AWS.

I put the data source in a new data-sources directory. I could place these in
the resources directory if you like.

I've also added a method to deploy the provider to the local provider to the
Makefile.