Respect passed configuration from terraform provider in the go SDK (#6)
* Respect passed configuration from terraform provider in the go SDK

Previously, the full support for definining properties like the cdp profile, acceess key id and
private key was defined, but the CDP Go SDK did not respect those. Now, we have support for passing
down all three provider defined configuration like this:
```
provider "cdp" {
  cdp_access_key_id = ""
  cdp_private_key = ""
  cdp_profile = "default"
}
```

profile got renamed to cdp_profile to be in line with the rest (and env variables).

Also added IAM client (for the Altus endpoint vs CDP endpoint) and fixed error handling in the
provider code so that it does not panic.