Some resource remodelling: resource_<service>_aws_<type>. (#5)
* Some resource remodelling: resource_<service>_aws_<type>.

This moves from using "cloud_platform" as a parameter to making resources AWS
specific. This is quite mechanical.

In my next change, I'll be reviewing the resources to make their models more
accurately reflect the actual create API call request naming and structure.

* Large rework of resource modelling.

This change pushes on better modelling of resources to match the API.

Note that the datahub cluster doesn't work yet. I'll work on that while I get a
review for this.

Note that I've added the ID broker mappings resource, moving this out of the
environment. I think this is right, but am not 100% confident yet.

Note I've added a NOTES.md file to track items and questions while we work on
this collaboratively (I may vanish at any point and don't want this lost). This
will be removed in the relatively short term as we get to an initial release and
move to more regular development.

I've also added a RESOURCES.md to clearly communicate the status of the various
resources and data sources. Again, this will go away shortly when we have real
documentation in place and when we get to an intial usable point.