Get the AWS datahub cluster resource working in a simple case. (#9)
* Get the AWS datahub cluster resource working in a simple case.

This change is enough to get the AWS datahub cluster working in at least one
simple case.

Note that I ran into an issue with the SDK's handling of optional booleans. It
does not send over the boolean value -- it sends null -- because of omitempty,
and this causes our backend to explode. It is reasonable to call this a bug in
the server, but in the short term I've adjust things to work around it.

* Fix AWS datahub cluster read.