Add IAM group resource, data source. A few cross-cutting things. (#13)
* Add IAM group resource, data source. A few cross-cutting things.

This change adds an IAM group resource and data source and let me try out doing
a change where I updated docs and change log inline.

As part of the above, I moved data source code into the resources directory. I
think this is more natural, and matches what I see in the AWS provider.

I bumped the hash of the service models in the SDK to pick up the addition of
x-nullable on a few optional, boolean properties. This was needed to get the
SDK to treat booleans as pointers and avoid swagger-go omitempty behavior. This
required changes in the dathub cluster resource.

I fixed an issue with endpoint construction for Altus style services to get the
calls to IAM to work.

While I was there, I added ML to the client object. I'm going to add the ML
workspace resource next.

I removed the main implementations of both the SDK and the provider. I think
this wasn't needed for anything.

* Respond to Enis's feedback