Add retries to cdp_aws_environment resource to ride over consistency issues (#20)
This adds retries using the terraform's resource.Retry() functionality. I added this only to the
aws_environment for now since this is the first resource I run into the issue. Will need to expand
coverage to the other resources, and find a common way to do the retries across all request types
(delete, read, etc).

WE parse the error codes returned from the API layer to check whether it is a specific type of
exception to decide whether to retry or not.

Testing done:
 - Tested manually by creating aws_environment resoureces in CDP.