Fix eventual consistency issue with AWS cross-account credential creation (#24)
In testing with the terraform-cdp-modules which creates a cross-account AWS role before the CDP
credential, there seems to be an eventual consistency issue from AWS IAM. This fix just adds
retry mechanism for the operation for 30 seconds before failing over. I am tryin to establish
this type of retrying behaviour everywhere to be more resilient across ephemeral issues.