Update CDP API spec files (#42)
I updated the CDP API spec files which were a couple of years old. We need a new API response field
for the listCredentials API. I've added some simple documentation for how to generate the swagger
code and how to update the API Spec files.

The other changes are simply comnig from changes in the API / swagger version.