Resource aws credential (#43)
* Rebase cdp_environments_aws_credential on top of the new terraform plugin framework.

This is the first resource that we are migrating from the SDKv2 to the pluging framework. Most of
these changes should be straighforward changes to the structure and not logic.

One big difference is that we do not handle retries now (framework does not support it). We'll
find a way to do retries (maybe using https://pkg.go.dev/github.com/hashicorp/go-retryablehttp
in a follow up PR.

Testing done:
 - No unit tests yet. I have tested this with a modified examples/credential module.

* Fux type in util function name

Co-authored-by: David Szabo <dszabo@cloudera.com>