Introduce cdp.Logger so that we can bridge CDP SDK logs to TF logs (#46)
I've added a simple interface cdp.Logger to the SDK with a default implementation that passes the
logs to the default golang logger. I've added the Context and Logger to the Config object so that
if they are passed, then we use those in the SDK.

The TFLoggerAdaptor is the bridge between the CDP logs to the TF logs. TF really wants the logs to
use the request Context otherwise the logs go to a blackhole. So we need to be careful in passing
the Context from Terraform request to the provider to the request to the SDK so that the SDK
request can have the terraform context as its parent context. The changes in the two resources that
we support are doing exactly that. All new data sources and resources we need to do the same thing.

I've had to migrate the authn package since golang packages cannot have a cyclic dependency. The
authn package did not make sense anyways and having everything under the cdp package makese it
simpler.

To see the logs at the debug level from the SDK, we can now run:
```
TF_LOG=DEBUG terraform apply
```

which then outputs requests to the CDP URLs:
```
2023-04-19T17:32:29.956-0700 [DEBUG] provider.terraform-provider-cdp: HTTP Request URL=https://$DOMAIN/api/v1/environments2/listCredentials method=POST status=200  resp=&{200 OK 200 HTTP/1.1 1 1 map[Connection:[keep-alive] Content-Length:[356] Content-Type:[application/json] Date:[Thu, 20 Apr 2023 00:32:29 GMT] X-Altus-Request-Id:[4ff6c6df-9cd5-469f-85f9-db9356f705a4] X-Cdp-Request-Id:[4ff6c6df-9cd5-469f-85f9-db9356f705a4]] 0x1400034fec0 356 [] false false map[] 0x140005fc100 0x140005d2630} durationMs=541873583: tf_rpc=ReadResource @module=cdp tf_provider_addr=terraform.cloudera.com/cloudera/cdp tf_req_id=a9cde6f9-6b72-df9d-282e-9846c34b10aa tf_resource_type=cdp_environments_aws_credential @caller=/Users/enis/projects/terraform-provider-cdp/provider/provider.go:201 timestamp=2023-04-19T17:32:29.955-0700
```