Documentation generation using tfplugindocs (#50)
I was able to get tfplugindocs working so that we can generate the resource and data-source
documentation partially from the existing Schema and examples from the code. This seems to be
best practice for first-party HashiCorp providers. This saves us from duplicating our work to
document each and every attribute in the resource docs seperately.

I have copied the docs/ content under templates/ and only changed the index.md for the provider
top-level page. I did not yet touch the docs for other resources and data-sources to keep the
changes in this PR to be smaller. We will do follow up PRs to migrate them to the new way.

I've also added examples under examples/provider and  sourced them from the templates/index.md.tmpl
so that these terraform files can also be used stand alone. We will have to do similar setup for
other resources and data-sources as well.

The go mod dependency was not working for me, so I moved utils/tfplugindocs.go to
tools/tfplugindocs.go.

I added some docuumentation about generating docs.