Add a pre commit test github workflow
I added a github workflow that uses github actions to run the build and the unit and terraform acceptance tests for every pull request. 

The contents of the github workflow mimics the https://github.com/hashicorp/terraform-provider-scaffolding-framework/tree/main/.github/workflows as well as other similar providers from Hashicorp. This seems like a good start for the initial test, but we will add more workflows (especially for releasing later).