CDPCP-9226 Improvements to docs (#16)
* CDPCP-9226 docs improvements

* Add .tmpl files for all resources and data-sources

I added a `.tmpl` file for every resource and data-source which is technically not needed since
`tfplugindocs` already generates the `.md` files under `docs/`. However there is no way to set
the subcategory in the fully-generated way, so we set it in the `.tmpl` files instead. Most of the
`.tmpl` files are identical, but we can add more content in specific resources when needed.

For all new resources and data-sources we should make sure that we add a similar file like this
```
templates/resources/<resource-name>.md.tmpl
```

and then execute:
```
make docs
```

As a part of this change, I have updated the CHANGELOG with the v0.1.0-rc2 contents as well.

* Formatted terraform files and regenerated docs

---------

Co-authored-by: Enis Soztutar <enis@cloudera.com>