Add a pre commit test github workflow (#14)
* Add a pre commit test github workflow

I added a github workflow that uses github actions to run the build and the unit and terraform acceptance tests for every pull request. 

The contents of the github workflow mimics the https://github.com/hashicorp/terraform-provider-scaffolding-framework/tree/main/.github/workflows as well as other similar providers from Hashicorp. This seems like a good start for the initial test, but we will add more workflows (especially for releasing later).

* Add Makefile build target

* Clean up golint issues

* Increase the linter timeout to 3mins

* Test regex matching for newlines on Windows

---------

Co-authored-by: Enis Soztutar <enis@cloudera.com>