Set User-Agent and x-altus-client-app headers for the golang SDK (#13)
* Set User-Agent and x-altus-client-app headers for the golang SDK

I've extended the SDK to be able to accept a custom User-Agent header as well as the header for
x-altus-client-app which is defined in the Java SDK (not sure whether it is used or not). The
user agent header is set to follow the pattern for CDP CLI and Java SDK.

Further, the terraform provider now overrides the user agent from the SDK with a Terraform specific
one. We intend to split the cdp-sdk-go into its own repository so once that happens, the user agent
from the go SDK vs the user agent from the terraform provider should be tracked differently.

"This commit does not contain secrets"

Testing done:
 - new unit tests

* Add one more unit test

* CDPCP-9682 Fixed build and test errors

---------

Co-authored-by: Enis Soztutar <enis@cloudera.com>
Co-authored-by: David Szabo <dszabo@cloudera.com>