CDPCP-9951 Clone swagger files locally
I added a clone-swaggers Makefile target to copy the swagger files from the upstream repo and added
them to this git repo. We can now reflect upstream changes to the local copies of these files (by
sending a PR against the swaggeers) before waiting on the public API release which can take 2-3
weeks. When the API changes are released upstream, changing the target API release version in
Makefile and running:
```
  make clone-swaggers
```
will just override the existing local copies with the upstream files so that we can keep the
swaggers in sync and will not diverge much.

Added more details in the README.