CDPCP-9511 Ability to set CDP Region (#32)
CDP Control plane has multiple regions, but we lacked the ability to set the region to send the
API requests to in the TF provider. We were defaulting to the us-west-1 region of the control plane.

With this change, I have added the cdp_region attribute to the provider configuration and support
for the same in the cdp-sdk-go. Similar to CDP CLI and Java SDK, we support setting the region
either from the provider, or from env variables or from the CDP config file. The changes to the
docs hopefully make it clear for the end user to configure this.

I have also done some light cleanup to the tests and reformatted the tf code and regenerated docs.

Testing done:
 - New and existing unit tests
 - Tested manually by setting the region to eu-1 in my local profile

Co-authored-by: Enis Soztutar <enis@cloudera.com>