Fixed issue caused by config names that have uppercase characters (#37)
* Fixed issue caused by config names that have uppercase characters

* Fixed unit test and add more test cases

We decided to follow the Python CDP CLI semantics for case sensitivity to be compatible between
the two tools and least surprises to the end user. Python INI file parsing defaults to case
sensitive section names, but case insensitive option names. We do the same in the go SDK.

This commit adds some more unit tests.

---------

Co-authored-by: Enis Soztutar <enis@cloudera.com>