Change the provider addresss to simply be cloudera/cdp (#35)
* Change the provider addresss to simply be cloudera/cdp

The provider address in the documentation does not have to be "fully qualified" with the terraform
registry domain. I have checked similar providers, none of them contain it. It is more readable
this way.

I have also removed the required providers configuration from the resource examples since I think
it is not needed. Also very small doc fixes.

Testing done:
 - doc only change

* Added a few missed resources since they were committed after this PR was created