Run Terraform fmt for examples files during the build
We disabled running `terraform fmt` on the `examples/` directory since the CI builds did not
have terraform in the container / VM image. Now that we are building with github, we can enable
that again both for local builds as well as builds for github checks. The only downside is that
all local builds will require terraform to be installed.