DWX-19297 Extend hive virtual warehouse with all available properties (#186)
This change exposes all properties required for creating a Hive virtual warehouse in a data warehouse CDP cluster. In addition to creation, it introduces the capability to delete the virtual warehouse. However, update and import functionalities are not included in this change. The virtual warehouse is dependent on the data warehouse cluster and a data lake object. Conflicting properties have been validated using the validation framework.
Testing

The changes have been tested using new unit and acceptance tests:
   - The acceptance tests create a CDP environment, a data lake, and an AWS-based Kubernetes cluster for the data warehouse. The Hive virtual warehouse is deployed on this Kubernetes cluster.
   - Both failure and success scenarios have been verified with the new unit tests.

Examples and documentation have been generated for this change.