CDPCP-14163 Increase hive auto suspend timeout
During e2e testing the hive warehouse went into a Stopping state
from the Starting state. Probably during the 30s of poll time the
auto suspend timeout was hit and that prevented us to see the
Running state of the entity. The auto suspend timeout will be set
to 20 minutes and that should be enough time for a warehouse to
start and the check to succeed.